Move tests to test package.
