+++ alitheia/core/src/test/java/eu/sqooss/service/admin/actions/test/TestAddProject.java
@@ -1,4 +1,4 @@
-package eu.sqooss.service.admin.actions;
+package eu.sqooss.service.admin.actions.test;
 
 import static org.junit.Assert.*;
 import static org.hamcrest.CoreMatchers.*;
